+++ common-custom-user-data-maven-extension/src/main/java/com/gradle/CustomBuildCacheConfig.java
@@ -1,0 +1,14 @@
+package com.gradle;
+
+import com.gradle.maven.extension.api.cache.BuildCacheApi;
+
+final class CustomBuildCacheConfig {
+
+    static void configureBuildCache(BuildCacheApi buildCache) {
+        // custom config goes here
+    }
+
+    private CustomBuildCacheConfig() {
+    }
+
+}